Move StatusLED class to new components module
